feat(snarker/ext): move worker stuff to a module
