+++ cli/src/commands/snarker/mod.rs
@@ -1,28 +1,18 @@
 use std::collections::BTreeMap;
-use std::ffi::{OsStr, OsString};
-use std::io;
-use std::mem::size_of;
-use std::process::Stdio;
+use std::ffi::OsString;
+
 use std::sync::Arc;
 use std::time::Duration;
 
-use binprot::macros::BinProtRead;
-use binprot::{BinProtRead, BinProtWrite};
 use mina_p2p_messages::v2::{
-    CurrencyFeeStableV1, NonZeroCurvePoint, SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponse,
-    SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponseA0, TransactionSnarkWorkTStableV2Proofs,
-    UnsignedExtendedUInt64Int64ForVersionTagsStableV1,
+    CurrencyFeeStableV1, UnsignedExtendedUInt64Int64ForVersionTagsStableV1,
 };
 use rand::prelude::*;
 use serde::Serialize;
 use shared::log::inner::Level;
-use snarker::external_snark_worker::{
-    ExternalSnarkWorkerError, ExternalSnarkWorkerEvent, ExternalSnarkWorkerService,
-    ExternalSnarkWorkerStartAction, ExternalSnarkWorkerWorkError, SnarkWorkSpec,
-};
+use snarker::external_snark_worker::ExternalSnarkWorkerStartAction;
 use snarker::p2p::service_impl::TaskSpawner;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
-use tokio::process::Command;
+
 use tokio::select;
 use tokio::sync::{mpsc, oneshot};
 
@@ -54,6 +44,8 @@ mod http_server;
 mod rpc;
 use rpc::RpcP2pConnectionOutgoingResponse;
 mod tracing;
+
+mod ext_snark_worker;
 
 /// Openmina snarker
 #[derive(Debug, clap::Args)]
@@ -336,7 +328,7 @@ struct SnarkerService {
     libp2p: Libp2pService,
     rpc: rpc::RpcService,
     stats: Stats,
-    snark_worker_sender: Option<ExternalSnarkWorkerSender>,
+    snark_worker_sender: Option<ext_snark_worker::ExternalSnarkWorkerSender>,
 }
 
 impl snarker::ledger::LedgerService for SnarkerService {
@@ -545,391 +537,5 @@ impl TaskSpawner for P2pTaskSpawner {
                 local_set.block_on(&runtime, fut);
             })
             .unwrap();
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum SnarkerError {
-    #[error(transparent)]
-    BinprotError(#[from] binprot::Error),
-    #[error(transparent)]
-    IOError(#[from] io::Error),
-    #[error(transparent)]
-    NixError(#[from] nix::Error),
-    #[error("external snark worker is not running")]
-    NotRunning,
-    #[error("external snark worker is busy")]
-    Busy,
-    /// Protocol logic is broken
-    #[error("communication is broken: {_0}")]
-    Broken(String),
-}
-
-impl From<SnarkerError> for ExternalSnarkWorkerError {
-    fn from(source: SnarkerError) -> Self {
-        match source {
-            SnarkerError::BinprotError(err) => {
-                ExternalSnarkWorkerError::BinprotError(err.to_string())
-            }
-            SnarkerError::IOError(err) => ExternalSnarkWorkerError::IOError(err.to_string()),
-            SnarkerError::NixError(err) => {
-                ExternalSnarkWorkerError::Error(format!("nix error: {err}"))
-            }
-            SnarkerError::NotRunning => ExternalSnarkWorkerError::NotRunning,
-            SnarkerError::Busy => ExternalSnarkWorkerError::Busy,
-            SnarkerError::Broken(err) => ExternalSnarkWorkerError::Broken(err),
-        }
-    }
-}
-
-impl From<SnarkerError> for ExternalSnarkWorkerEvent {
-    fn from(source: SnarkerError) -> Self {
-        ExternalSnarkWorkerEvent::Error(source.into())
-    }
-}
-
-async fn write_snark_work_spec<T: BinProtWrite, W: AsyncWrite + Unpin>(
-    spec: T,
-    mut w: W,
-) -> Result<(), SnarkerError> {
-    let mut buf = Vec::new();
-    spec.binprot_write(&mut buf)?;
-    let len = (buf.len() as u64).to_le_bytes();
-    w.write_all(&len).await?;
-    w.write_all(&buf).await?;
-    Ok(())
-}
-
-#[derive(BinProtRead)]
-pub enum ExternalSnarkWorkerResult {
-    Ok(Option<TransactionSnarkWorkTStableV2Proofs>),
-    Err(String),
-}
-
-async fn read_snark_work_result<R: AsyncRead + Unpin>(
-    mut r: R,
-) -> Result<ExternalSnarkWorkerResult, SnarkerError> {
-    let mut len_buf = [0; size_of::<u64>()];
-    r.read_exact(&mut len_buf).await?;
-    let len = u64::from_le_bytes(len_buf);
-
-    let mut buf = Vec::with_capacity(len as usize);
-    let mut r = r.take(len);
-    r.read_to_end(&mut buf).await?;
-
-    let mut read = buf.as_slice();
-    let result = ExternalSnarkWorkerResult::binprot_read(&mut read)?;
-    Ok(result)
-}
-
-struct ExternalSnarkWorkerSender {
-    data_chan: mpsc::Sender<SnarkWorkSpec>,
-    cancel_chan: mpsc::Sender<()>,
-    kill_chan: oneshot::Sender<()>,
-}
-
-impl ExternalSnarkWorkerSender {
-    fn start<P: AsRef<OsStr>>(
-        path: P,
-        public_key: NonZeroCurvePoint,
-        fee: CurrencyFeeStableV1,
-        event_sender: mpsc::UnboundedSender<Event>,
-    ) -> Result<Self, SnarkerError> {
-        let (data_chan, mut data_rx) = mpsc::channel(1);
-        let (cancel_chan, mut cancel_rx) = mpsc::channel(1);
-        let (kill_chan, kill_rx) = oneshot::channel();
-
-        let mut cmd = Command::new(path);
-
-        // TODO(akoptelov) make the block return terminal errors instead of sending them down the channel and exit.
-        std::thread::Builder::new()
-            .name("external-snark-worker".into())
-            .spawn(move || {
-                let runtime = tokio::runtime::Builder::new_current_thread()
-                    .enable_all()
-                    .build()
-                    .unwrap();
-                runtime.block_on(async move {
-                    let event_sender_clone = event_sender.clone();
-                    let event = move |event: ExternalSnarkWorkerEvent| {
-                        if let Err(err) = event_sender_clone.send(event.into()) {
-                            eprintln!("error sending event: {err}");
-                        }
-                    };
-
-                    let mut child = match cmd
-                        .stdin(Stdio::piped())
-                        .stdout(Stdio::piped())
-                        .stderr(Stdio::inherit())
-                        .spawn()
-                    {
-                        Ok(v) => v,
-                        Err(err) => {
-                            event(SnarkerError::from(err).into());
-                            return;
-                        }
-                    };
-
-                    event(ExternalSnarkWorkerEvent::Started);
-
-                    let mut child_stdin = child.stdin.take().unwrap();
-                    let mut child_stdout = child.stdout.take().unwrap();
-
-                    if let Some(pid) = child.id() {
-                        let pid = nix::unistd::Pid::from_raw(pid as i32);
-                        tokio::spawn(async move {
-                            loop {
-                                let Some(spec) = data_rx.recv().await else {
-                                    return;
-                                };
-                                let spec = SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponse(Some((
-                                    SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponseA0 {
-                                        instances: spec,
-                                        fee: fee.clone(),
-                                    },
-                                    public_key.clone(),
-                                )));
-
-                                if let Err(err) = write_snark_work_spec(spec, &mut child_stdin).await {
-                                    event(err.into());
-                                    return;
-                                }
-                                let response = read_snark_work_result(&mut child_stdout).await;
-                                match response {
-                                    Ok(result) => {
-                                        match result {
-                                            ExternalSnarkWorkerResult::Ok(Some(v)) => {
-                                                event(Arc::new(v).into());
-                                            }
-                                            ExternalSnarkWorkerResult::Ok(None) => {
-                                                event(ExternalSnarkWorkerEvent::WorkCancelled);
-                                            }
-                                            ExternalSnarkWorkerResult::Err(err) => {
-                                                event(ExternalSnarkWorkerWorkError::Error(err).into());
-                                            }
-                                        }
-                                    }
-                                    Err(err) => {
-                                        let _ = event(err.into());
-                                    }
-                                }
-                            }
-                        });
-
-                        let event_sender_clone = event_sender.clone();
-                        tokio::spawn(async move {
-                            loop {
-                                if cancel_rx.recv().await.is_none() {
-                                    return;
-                                }
-                                println!("sending cancel signal to {pid}...");
-                                if let Err(err) = nix::sys::signal::kill(pid, nix::sys::signal::Signal::SIGINT) {
-                                    let _ = event_sender_clone.send(ExternalSnarkWorkerEvent::Error(SnarkerError::from(err).into()).into());
-                                    return;
-                                }
-                            }
-                        });
-
-                        tokio::select! {
-                            _ = kill_rx => {
-                                if let Err(err) = child.kill().await {
-                                    let _ = event_sender.send(ExternalSnarkWorkerEvent::Error(SnarkerError::from(err).into()).into());
-                                } else {
-                                    let _ = event_sender.send(ExternalSnarkWorkerEvent::Killed.into());
-                                }
-                                return;
-                            }
-                            _ = child.wait() => {
-                                return
-                            }
-                        };
-                    }
-                });
-            })?;
-
-        Ok(ExternalSnarkWorkerSender {
-            data_chan,
-            cancel_chan,
-            kill_chan,
-        })
-    }
-
-    fn cancel(&mut self) -> Result<(), SnarkerError> {
-        self.cancel_chan
-            .try_send(())
-            .map_err(|_| SnarkerError::Broken("already cancelled".into()))
-    }
-
-    fn submit(&mut self, spec: SnarkWorkSpec) -> Result<(), SnarkerError> {
-        self.data_chan
-            .try_send(spec)
-            .map_err(|_| SnarkerError::Busy)
-    }
-
-    fn kill(self) -> Result<(), SnarkerError> {
-        self.kill_chan
-            .send(())
-            .map_err(|_| SnarkerError::Broken("already sent kill".into()))
-    }
-}
-
-impl ExternalSnarkWorkerService for SnarkerService {
-    fn start<P: AsRef<OsStr>>(
-        &mut self,
-        path: P,
-        public_key: NonZeroCurvePoint,
-        fee: CurrencyFeeStableV1,
-    ) -> Result<(), snarker::external_snark_worker::ExternalSnarkWorkerError> {
-        let cmd_sender =
-            ExternalSnarkWorkerSender::start(path, public_key, fee, self.event_sender.clone())?;
-        self.snark_worker_sender = Some(cmd_sender);
-        Ok(())
-    }
-
-    fn submit(
-        &mut self,
-        spec: SnarkWorkSpec,
-    ) -> Result<(), snarker::external_snark_worker::ExternalSnarkWorkerError> {
-        self.snark_worker_sender
-            .as_mut()
-            .ok_or(SnarkerError::NotRunning)
-            .and_then(|sender| sender.submit(spec))?;
-        Ok(())
-    }
-
-    fn cancel(&mut self) -> Result<(), ExternalSnarkWorkerError> {
-        self.snark_worker_sender
-            .as_mut()
-            .ok_or(SnarkerError::NotRunning)
-            .and_then(|sender| sender.cancel())?;
-        Ok(())
-    }
-
-    fn kill(&mut self) -> Result<(), snarker::external_snark_worker::ExternalSnarkWorkerError> {
-        self.snark_worker_sender
-            .take()
-            .ok_or(SnarkerError::NotRunning)
-            .and_then(|sender| sender.kill())?;
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod external_snark_worker_tests {
-    use std::{env, ffi::OsString, path::Path, time::Duration};
-
-    use binprot::BinProtRead;
-    use mina_p2p_messages::v2::{
-        CurrencyFeeStableV1, NonZeroCurvePoint, SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponse,
-        SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponseA0,
-    };
-    use snarker::{
-        event_source::Event,
-        external_snark_worker::{ExternalSnarkWorkerEvent, SnarkWorkSpec},
-    };
-    use tokio::sync::mpsc;
-
-    use crate::commands::snarker::ExternalSnarkWorkerSender;
-
-    macro_rules! expect_event {
-        ($source:expr, $event:pat) => {
-            let result = $source.recv().await.expect("failed to receive an event");
-            let Event::ExternalSnarkWorker(result) = result else {
-                panic!("unexpected event kind");
-            };
-            let $event = result else {
-                panic!("unexpected snark worker event: {result:?}");
-            };
-        };
-    }
-
-    fn mina_exe_path() -> OsString {
-        env::var_os("MINA_EXE_PATH")
-            .or_else(|| {
-                env::var_os("CARGO_MANIFEST_DIR")
-                    .map(|dir| Path::new(&dir).join("bin/mina").into_os_string())
-            })
-            .unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_kill() {
-        let (event_tx, mut event_rx) = mpsc::unbounded_channel();
-        let cmd_sender = ExternalSnarkWorkerSender::start(
-            mina_exe_path(),
-            NonZeroCurvePoint::default(),
-            CurrencyFeeStableV1(
-                mina_p2p_messages::v2::UnsignedExtendedUInt64Int64ForVersionTagsStableV1(
-                    10_i64.into(),
-                ),
-            ),
-            event_tx,
-        )
-        .unwrap();
-
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Started);
-
-        cmd_sender.kill().expect("cannot kill worker");
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Killed);
-    }
-
-    fn read_input<R: std::io::Read>(
-        mut r: R,
-    ) -> (NonZeroCurvePoint, CurrencyFeeStableV1, SnarkWorkSpec) {
-        let SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponse(Some((
-            SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponseA0 { instances, fee },
-            public_key,
-        ))) = SnarkWorkerWorkerRpcsVersionedGetWorkV2TResponse::binprot_read(&mut r)
-            .expect("cannot read work spec")
-        else {
-            unreachable!("incorrect work spec");
-        };
-
-        (public_key, fee, instances)
-    }
-
-    #[tokio::test]
-    async fn test_work() {
-        const DATA: &[u8] = include_bytes!("../../../tests/files/snark_spec/spec1.bin");
-        let mut r = DATA;
-        let (public_key, fee, instances) = read_input(&mut r);
-
-        let (event_tx, mut event_rx) = mpsc::unbounded_channel();
-        let mut cmd_sender =
-            ExternalSnarkWorkerSender::start(mina_exe_path(), public_key, fee, event_tx).unwrap();
-
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Started);
-
-        cmd_sender.submit(instances).unwrap();
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::WorkResult(_));
-
-        cmd_sender.kill().expect("cannot kill worker");
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Killed);
-    }
-
-    #[tokio::test]
-    async fn test_cancel() {
-        const DATA: &[u8] = include_bytes!("../../../tests/files/snark_spec/spec1.bin");
-        let mut r = DATA;
-        let (public_key, fee, instances) = read_input(&mut r);
-
-        let (event_tx, mut event_rx) = mpsc::unbounded_channel();
-        let mut cmd_sender =
-            ExternalSnarkWorkerSender::start(mina_exe_path(), public_key, fee, event_tx).unwrap();
-
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Started);
-
-        cmd_sender.submit(instances.clone()).unwrap();
-
-        tokio::time::sleep(Duration::from_secs(60)).await;
-
-        cmd_sender.cancel().unwrap();
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::WorkCancelled);
-
-        cmd_sender.submit(instances).unwrap();
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::WorkResult(_));
-
-        cmd_sender.kill().expect("cannot kill worker");
-        expect_event!(event_rx, ExternalSnarkWorkerEvent::Killed);
     }
 }
